feat(FileTransferClient): implement automatic file rename mechanism
When a file is transfered, the output name is set to a temporary random name while it is being transfered.
After the transfer is done, the output name is set to the original file name.
If file name collisions occur, an increasing number suffix will be appended at the output name.